+++ src/dfx/src/config/dfinity.rs
@@ -164,9 +164,9 @@ pub struct ConfigCanistersCanister {
 
     /// # Shrink Canister WASM
     /// Whether run `ic-wasm shrink` after building the Canister.
-    /// Default is true.
-    #[serde(default = "default_as_true")]
-    pub shrink: bool,
+    /// Enabled by default for Rust/Motoko canisters.
+    /// Disabled by default for custom canisters.
+    pub shrink: Option<bool>,
 
     /// # Metadata
     /// Defines metadata sections to set in the canister .wasm