+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # UNRELEASED
 
+## DFX
+
+### fix: default not shrink for custom canisters
+
 ## Dependencies
 
 ### Replica