+++ CHANGELOG.md
@@ -4,6 +4,27 @@
 
 ## DFX
 
+### fix: default not shrink for custom canisters
+
+## Dependencies
+
+### Replica
+
+Updated replica to elected commit dcbf401f27d9b48354e68389c6d8293c4233b055.
+This incorporates the following executed proposals:
+
+- [90485](https://dashboard.internetcomputer.org/proposal/90485)
+- [90008](https://dashboard.internetcomputer.org/proposal/90008)
+
+### Frontend canister
+
+- Module hash: db07e7e24f6f8ddf53c33a610713259a7c1eb71c270b819ebd311e2d223267f0
+- https://github.com/dfinity/sdk/pull/2753
+
+# 0.12.0
+
+## DFX
+
 ### feat(frontend-canister): add warning if config is provided in `.ic-assets.json` but not used
 
 ### fix(frontend-canister): Allow overwriting default HTTP Headers for assets in frontend canister